Dashboard panel was fixed
