+++ src/containers/dashboard/BalanceInfo/styles.css
@@ -1,39 +1,23 @@
-.button {
-  margin-bottom: 15px;
+.balance {
+  width: 350px;
+  background-color: #394b59;
+  border-radius: 3px;
+  box-shadow: 0 0 0 1px rgba(16, 22, 26, 0.4);
+  padding: 20px;
+  margin-bottom: 20px;
 }
 
-.block {
-  padding: 15px 0;
-  border-bottom: 1px solid #e6e8eb;
+.button {
+  width: 121px;
 }
 
 .val {
-  font: normal 32px Roboto;
-  color: #222;
+  font: normal 27px Roboto;
+  color: #f5f8fa;
 }
 
 .label {
-  font: normal 16px Roboto;
-  color: #6e7987;
-  margin-top: 5px;
-}
-
-.currencybutton {
-  font: normal 16px Roboto;
-  color: #0080ff;
-  background-color: transparent;
-  outline: none;
-  margin: 0;
-  padding: 0;
-  cursor: pointer;
-  display: inline;
-  border: none;
-}
-
-.activeCurrencyButton {
-  composes: currencybutton;
-  pointer-events: none;
-  cursor: default;
-  font: bold 16px Roboto;
-  color: #6e7987;
+  font: normal 14px Roboto;
+  color: #bfccd6;
+  margin: 7px 0 25px;
 }
