+++ src/containers/dashboard/Dashboard/styles.css
@@ -5,7 +5,7 @@
 }
 
 .main {
-  width: calc(100% - 250px);
+  width: calc(100% - 350px);
   height: 100%;
   float: left;
   vertical-align: top;
