+++ src/containers/dashboard/IcoInfo/styles.scss
@@ -1,0 +1,42 @@
+.container {
+  width: 350px;
+  background-color: #394b59;
+  border-radius: 3px;
+  box-shadow: 0 0 0 1px rgba(16, 22, 26, 0.4);
+  padding: 20px;
+}
+
+.button {
+  margin-bottom: 15px;
+}
+
+.val {
+  font: normal 27px Roboto;
+  color: #f5f8fa;
+}
+
+.label {
+  font: normal 14px Roboto;
+  color: #bfccd6;
+  margin: 7px 0 25px;
+}
+
+.currencybutton {
+  font: normal 16px Roboto;
+  color: #0080ff;
+  background-color: transparent;
+  outline: none;
+  margin: 0;
+  padding: 0;
+  cursor: pointer;
+  display: inline;
+  border: none;
+}
+
+.activeCurrencyButton {
+  composes: currencybutton;
+  pointer-events: none;
+  cursor: default;
+  font: bold 16px Roboto;
+  color: #6e7987;
+}
