+++ src/locales/en/dashboard.json
@@ -28,8 +28,8 @@
   "tokensSold": "Tokens sold",
   "raisedIn": "Raised in",
   "daysToGo": "Days to go",
-  "purchaseTokens": "Purchase tokens",
-  "buyTokens": "Buy Tokens",
+  "purchaseTokens": "Contribute",
+  "buyTokens": "Buy ERC20 tokens",
   "gasFee": "Gas fee:",
   "minContribution": "Min. contribution:",
   "contributeAll": "Contribute all",
